the rating ranges are updated from 0..100 to 0..10

+++ database/migrations/2019_10_22_094957_create_reviews_table.php
@@ -16,7 +16,7 @@ class CreateReviewsTable extends Migration
         Schema::create('reviews', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
-            $table->smallInteger('points')
+            $table->tinyInteger('points')
                   ->unsigned();
             $table->string('comment');
             $table->integer('user_id')
